Failed attempt to tests unix socket
I can't tests unix sockets. To do that I must create a Linux User in the
DB containers. From the makefile using the containers.podman collection
and community.docker for GHA. In order to use ansible against the
DB containers, I'll need to install Python there first. So, I stop
working on that. Too complicated for a bug that comes from MariaDB
10.6 only.